raised the buffer size from max 4gb

+++ include/renderer/resources/buffer.hpp
@@ -12,7 +12,7 @@ namespace gfx
 	{ 
 		gfx::BufferUsage usage = gfx::BufferUsage::UNIFORM;
 		gfx::BufferUsageHint hint = gfx::BufferUsageHint::STATIC;
-		uint32_t byteSize = 0;
+		uint64_t byteSize = 0;
 		utils::Span<uint8_t> initialData;
 	};
 }
